third_party/WebKit/LayoutTests/web-animations-api/animation-finish-event-cancelled.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Canceling an animation: finish event</title>
<link rel="help" href="https://w3c.github.io/web-animations/#canceling-an-animation-section">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<body>
<script>
var anim1 = document.body.animate([], 100000);
var anim2 = document.body.animate([], 100000);
var anim3 = document.body.animate([], 100000);

var noFinish = async_test('Animation finish event should not fire when cancelled');

anim1.onfinish = function() {
  noFinish.step(function() {
    anim3.finish();
  });
};

anim2.onfinish = function() {
  noFinish.step(function() {
    assert_true(false);
  });
};

anim3.onfinish = function() {
  noFinish.done();
};

anim1.finish();
anim2.cancel();
</script>
